Add variable list to main LinkModel instead of extending for Explore only
