+++ packages/grafana-runtime/src/services/templateSrv.ts
@@ -1,22 +1,4 @@
-import { ScopedVars, TimeRange, TypedVariableModel } from '@grafana/data';
-
-/**
- * Can be used to gain more information about an interpolation operation
- */
-export interface VariableInterpolation {
-  /** The full matched expression including, example: ${varName.field:regex} */
-  match: string;
-  /** In the expression ${varName.field:regex} variableName is varName */
-  variableName: string;
-  /** In the expression ${varName.fields[0].name:regex} the fieldPath is fields[0].name */
-  fieldPath?: string;
-  /** In the expression ${varName:regex} the regex part is the format */
-  format?: string;
-  /** The formatted value of the variable expresion. Will equal match when variable not found or scopedVar was undefined or null **/
-  value: string;
-  // When value === match this will be true, meaning the variable was not found
-  found?: boolean;
-}
+import { ScopedVars, TimeRange, TypedVariableModel, VariableInterpolation } from '@grafana/data';
 
 /**
  * Via the TemplateSrv consumers get access to all the available template variables
